Initial commit
This commit adds the bases. No useful code, only the structure.

Note: the structure is similar to Control's, with the app code in
src/, the JS in src/app/, the CSS in src/css/, libs in src/libs/.

The difference is that libs are not included in the repo and have
to be manually fetched with `npm install` and `bower install`.

This is to reduce the size of the repo and avoid making Git slower.