Reload config on `message::reloadConfig`
This commit adds an event listener on window `message` events, which
will check the origin before broadcasting an Angular event.

The event listener will also look for reloadConfig messages and
broadcast the appropriate event.

PlayerCtrl now listens for message::reloadConfig events and will
reinitialise with the new config when needed.