Fix image changes by using a fixed angular-preload-image
The original angular-preload-image did not handle ng-src changes at all
which meant that any changes to ng-src would not affect the src
attribute at all, unlike angular's default behaviour.

This is fixed in the fork (leoetlino/angular-preload-image on GitHub).
All this commit does is update bower.json to use the new dependency.