Player: Add a timeout timer to toggle()
This commit adds a timeout timer to toggle() so that it doesn't get
stuck in trying to play the stream. It will time out after 5 seconds
and reset the play/stop button state so that the user can try again.

This also handles the case where audio.paused is true even right after
calling audio.play(): this can happen on some mobile browsers, which
refuse the play the stream unless the action was initiated by the user.
We now check for this and will immediately abort the play attempt and
reset the play button so the user can play the stream normally, albeit
manually.

Fixes issue #8 (internal GitLab).